Move to book list dir
