+++ resources/js/components/BookListView/index.js
@@ -7,7 +7,7 @@ import { isEmpty } from "../../utils";
 
 //import { ConnectedGenres } from "../../containers";
 import Genres from '../Genres';
-import { Search } from "../Search";
+import { Search } from "./Search";
 import { Loading } from "../shared/Loading";
 import { ISBNModal } from "../shared/book/ISBNModal";
 import BooksSuspense from "./BooksSuspense";
